Count Numbers on fingers
If you start to count a number through your one hand fingers(5 fingers) , so if you start your count of 7 from little finger to thumb , you should go through all the fingers and reverse the path like 1 would be little finger and 5 would be thumb , the number 6 should be on the index finger and 7 on the middle finger. He asked me to write an algo so that whenever I give you a number you should be able to tell which finger that number would come to.