Create 3Sum.java
Given an array nums of n integers, are there elements a, b, c in nums such that a + b + c = 0?

Algorithm👨‍🎓
Create two Lists and one TreeSet.
Sort the given input array. using Arrays.sort() function.
Create a for loop that iterates from 0 to n-2.
Create two variables l and r. l is to iterate from the left and r is to iterate from the end.
The logic behind this approach is after we have sorted that data. At any moment of time if the triplet sum is greater than the target value.
Then we should search for the triplet before the given indices. Similarly, if the sum value is lesser than the target we should increase the iterator from the left.
Now if the triplet equals the sum. Add the elements in a list (row).📃
Add that particular row into the tree set. When we save it in tree set. if we use contains function it results in O(1). So that retrieval it is more efficient.
After looping ➰for all each iteration.
Return the Result.🔚