Create BottomViewOfBinaryTree.java
Given a Binary Tree, The task is to print the bottom view from left to right. A node x is there in output if x is the bottommost node at its horizontal distance. The horizontal distance of the left child of a node x is equal to a horizontal distance of x minus 1, and that of a right child is the horizontal distance of x plus 1. 

Reference : https://www.geeksforgeeks.org/bottom-view-binary-tree/